refactor: implement `controller.js` error handling abstraction
